+++ tests/e2e/api/src/models/products/index.ts
@@ -1,3 +1,5 @@
-export * from './abstract-product';
-export * from './shared-types';
+export * from './abstract';
+export * from './shared';
 export * from './simple-product';
+export * from './grouped-product';
+export * from './external-product';