+++ tests/e2e/api/src/repositories/rest/coupons/index.ts
@@ -1,0 +1,3 @@
+import couponRESTRepository from './coupon';
+
+export { couponRESTRepository };