+++ tests/e2e/api/src/repositories/rest/index.ts
@@ -1,2 +1,3 @@
 export * from './products';
 export * from './settings';
+export * from './coupons';