+++ tests/e2e/api/src/repositories/rest/products/shared.ts
@@ -1,22 +1,31 @@
 import {
+	AddPropertyTransformation,
+	CustomTransformation,
+	IgnorePropertyTransformation,
+	KeyChangeTransformation,
 	ModelTransformation,
 	ModelTransformer,
-	TransformationOrder,
-	KeyChangeTransformation,
-	AddPropertyTransformation,
-	IgnorePropertyTransformation,
+	ModelTransformerTransformation,
 	PropertyType,
 	PropertyTypeTransformation,
-	CustomTransformation,
-	ModelTransformerTransformation,
+	TransformationOrder,
 } from '../../../framework';
 import {
 	AbstractProduct,
+	IProductCrossSells,
+	IProductDelivery,
+	IProductExternal,
+	IProductGrouped,
+	IProductInventory,
+	IProductPrice,
+	IProductSalesTax,
+	IProductShipping,
+	IProductUpSells,
+	MetaData,
 	ProductAttribute,
 	ProductDownload,
 	ProductImage,
 	ProductTerm,
-	MetaData,
 } from '../../../models';
 import { createMetaDataTransformer } from '../shared';
 
@@ -122,16 +131,236 @@ export function createProductTransformer< T extends AbstractProduct >(
 			[
 				'date_created',
 				'date_modified',
-				'date_on_sale_from',
-				'date_on_sale_to',
 			],
 		),
 		new ModelTransformerTransformation( 'categories', ProductTerm, createProductTermTransformer() ),
 		new ModelTransformerTransformation( 'tags', ProductTerm, createProductTermTransformer() ),
 		new ModelTransformerTransformation( 'attributes', ProductAttribute, createProductAttributeTransformer() ),
 		new ModelTransformerTransformation( 'images', ProductImage, createProductImageTransformer() ),
-		new ModelTransformerTransformation( 'downloads', ProductDownload, createProductDownloadTransformer() ),
 		new ModelTransformerTransformation( 'metaData', MetaData, createMetaDataTransformer() ),
+		new PropertyTypeTransformation(
+			{
+				created: PropertyType.Date,
+				modified: PropertyType.Date,
+				isPurchasable: PropertyType.Boolean,
+				isFeatured: PropertyType.Boolean,
+				allowReviews: PropertyType.Boolean,
+				averageRating: PropertyType.Integer,
+				numRatings: PropertyType.Integer,
+				totalSales: PropertyType.Integer,
+				parentId: PropertyType.Integer,
+				menuOrder: PropertyType.Integer,
+				permalink: PropertyType.String,
+				relatedIds: PropertyType.Integer,
+			},
+		),
+		new KeyChangeTransformation< AbstractProduct >(
+			{
+				created: 'date_created_gmt',
+				modified: 'date_modified_gmt',
+				postStatus: 'status',
+				shortDescription: 'short_description',
+				isPurchasable: 'purchasable',
+				isFeatured: 'featured',
+				catalogVisibility: 'catalog_visibility',
+				allowReviews: 'reviews_allowed',
+				averageRating: 'average_rating',
+				numRatings: 'rating_count',
+				metaData: 'meta_data',
+				totalSales: 'total_sales',
+				parentId: 'parent_id',
+				menuOrder: 'menu_order',
+				relatedIds: 'related_ids',
+				links: '_links',
+			},
+		),
+	);
+
+	return new ModelTransformer( transformations );
+}
+
+/**
+ * Create a transformer for the product price properties.
+ */
+export function createProductPriceTransformation(): ModelTransformation[] {
+	const transformations = [
+		new IgnorePropertyTransformation(
+			[
+				'date_on_sale_from',
+				'date_on_sale_to',
+			],
+		),
+		new PropertyTypeTransformation(
+			{
+				onSale: PropertyType.Boolean,
+				saleStart: PropertyType.Date,
+				saleEnd: PropertyType.Date,
+				priceHtml: PropertyType.String,
+			},
+		),
+		new KeyChangeTransformation< IProductPrice >(
+			{
+				regularPrice: 'regular_price',
+				onSale: 'on_sale',
+				salePrice: 'sale_price',
+				saleStart: 'date_on_sale_from_gmt',
+				saleEnd: 'date_on_sale_to_gmt',
+				priceHtml: 'price_html',
+			},
+		),
+	];
+
+	return transformations;
+}
+
+/**
+ * Create a transformer for the product cross sells property.
+ */
+export function createProductCrossSellsTransformation(): ModelTransformation[] {
+	const transformations = [
+		new PropertyTypeTransformation(
+			{
+				crossSellIds: PropertyType.Integer,
+			},
+		),
+		new KeyChangeTransformation< IProductCrossSells >(
+			{
+				crossSellIds: 'cross_sell_ids',
+			},
+		),
+	];
+
+	return transformations;
+}
+
+/**
+ * Create a transformer for the product upsells property.
+ */
+export function createProductUpSellsTransformation(): ModelTransformation[] {
+	const transformations = [
+		new PropertyTypeTransformation(
+			{
+				upSellIds: PropertyType.Integer,
+			},
+		),
+		new KeyChangeTransformation< IProductUpSells >(
+			{
+				upSellIds: 'upsell_ids',
+			},
+		),
+	];
+
+	return transformations;
+}
+
+/**
+ * Transformer for the grouped products property.
+ */
+export function createProductGroupedTransformation(): ModelTransformation[] {
+	const transformations = [
+		new PropertyTypeTransformation(
+			{
+				groupedProducts: PropertyType.Integer,
+			},
+		),
+		new KeyChangeTransformation< IProductGrouped >(
+			{
+				groupedProducts: 'grouped_products',
+			},
+		),
+	];
+
+	return transformations;
+}
+
+/**
+ * Create a transformer for product delivery properties.
+ */
+export function createProductDeliveryTransformation(): ModelTransformation[] {
+	const transformations = [
+		new ModelTransformerTransformation( 'downloads', ProductDownload, createProductDownloadTransformer() ),
+		new PropertyTypeTransformation(
+			{
+				isVirtual: PropertyType.Boolean,
+				isDownloadable: PropertyType.Boolean,
+				downloadLimit: PropertyType.Integer,
+				daysToDownload: PropertyType.Integer,
+				purchaseNote: PropertyType.String,
+			},
+		),
+		new KeyChangeTransformation< IProductDelivery >(
+			{
+				isVirtual: 'virtual',
+				isDownloadable: 'downloadable',
+				downloadLimit: 'download_limit',
+				daysToDownload: 'download_expiry',
+				purchaseNote: 'purchase_note',
+			},
+		),
+	];
+
+	return transformations;
+}
+
+/**
+ * Create a transformer for product inventory properties.
+ */
+export function createProductInventoryTransformation(): ModelTransformation[] {
+	const transformations = [
+		new PropertyTypeTransformation(
+			{
+				trackInventory: PropertyType.Boolean,
+				remainingStock: PropertyType.Integer,
+				canBackorder: PropertyType.Boolean,
+				isOnBackorder: PropertyType.Boolean,
+				onePerOrder: PropertyType.Boolean,
+				stockStatus: PropertyType.String,
+				backOrderStatus: PropertyType.String,
+			},
+		),
+		new KeyChangeTransformation< IProductInventory >(
+			{
+				trackInventory: 'manage_stock',
+				remainingStock: 'stock_quantity',
+				stockStatus: 'stock_status',
+				onePerOrder: 'sold_individually',
+				backorderStatus: 'backorders',
+				canBackorder: 'backorders_allowed',
+				isOnBackorder: 'backordered',
+			},
+		),
+	];
+
+	return transformations;
+}
+
+/**
+ * Create a transformer for product sales tax properties.
+ */
+export function createProductSalesTaxTransformation(): ModelTransformation[] {
+	const transformations = [
+		new PropertyTypeTransformation(
+			{
+				taxClass: PropertyType.String,
+				taxStatus: PropertyType.String,
+			},
+		),
+		new KeyChangeTransformation< IProductSalesTax >(
+			{
+				taxStatus: 'tax_status',
+				taxClass: 'tax_class',
+			},
+		),
+	];
+
+	return transformations;
+}
+
+/**
+ * Create a transformer for product shipping properties.
+ */
+export function createProductShippingTransformation(): ModelTransformation[] {
+	const transformations = [
 		new CustomTransformation(
 			TransformationOrder.Normal,
 			( properties: any ) => {
@@ -163,66 +392,45 @@ export function createProductTransformer< T extends AbstractProduct >(
 		),
 		new PropertyTypeTransformation(
 			{
-				created: PropertyType.Date,
-				modified: PropertyType.Date,
-				isPurchasable: PropertyType.Boolean,
-				isFeatured: PropertyType.Boolean,
-				isVirtual: PropertyType.Boolean,
-				onePerOrder: PropertyType.Boolean,
-				onSale: PropertyType.Boolean,
-				saleStart: PropertyType.Date,
-				saleEnd: PropertyType.Date,
-				isDownloadable: PropertyType.Boolean,
-				downloadLimit: PropertyType.Integer,
-				daysToDownload: PropertyType.Integer,
 				requiresShipping: PropertyType.Boolean,
 				isShippingTaxable: PropertyType.Boolean,
-				trackInventory: PropertyType.Boolean,
-				remainingStock: PropertyType.Integer,
-				canBackorder: PropertyType.Boolean,
-				isOnBackorder: PropertyType.Boolean,
-				allowReviews: PropertyType.Boolean,
-				averageRating: PropertyType.Integer,
-				numRatings: PropertyType.Integer,
+				shippingClass: PropertyType.String,
+				shippingClassId: PropertyType.Integer,
+				weight: PropertyType.String,
 			},
 		),
-		new KeyChangeTransformation< AbstractProduct >(
+		new KeyChangeTransformation< IProductShipping >(
 			{
-				created: 'date_created_gmt',
-				modified: 'date_modified_gmt',
-				postStatus: 'status',
-				shortDescription: 'short_description',
-				isPurchasable: 'purchasable',
-				isFeatured: 'featured',
-				isVirtual: 'virtual',
-				catalogVisibility: 'catalog_visibility',
-				regularPrice: 'regular_price',
-				onePerOrder: 'sold_individually',
-				taxStatus: 'tax_status',
-				taxClass: 'tax_class',
-				onSale: 'on_sale',
-				salePrice: 'sale_price',
-				saleStart: 'date_on_sale_from_gmt',
-				saleEnd: 'date_on_sale_to_gmt',
-				isDownloadable: 'downloadable',
-				downloadLimit: 'download_limit',
-				daysToDownload: 'download_expiry',
 				requiresShipping: 'shipping_required',
 				isShippingTaxable: 'shipping_taxable',
 				shippingClass: 'shipping_class',
-				trackInventory: 'manage_stock',
-				remainingStock: 'stock_quantity',
-				stockStatus: 'stock_status',
-				backorderStatus: 'backorders',
-				canBackorder: 'backorders_allowed',
-				isOnBackorder: 'backordered',
-				allowReviews: 'reviews_allowed',
-				averageRating: 'average_rating',
-				numRatings: 'rating_count',
-				metaData: 'meta_data',
+				shippingClassId: 'shipping_class_id',
 			},
 		),
-	);
+	];
 
-	return new ModelTransformer( transformations );
+	return transformations;
 }
+
+/**
+ * Transformer for the properties unique to the external product type.
+ */
+export function createProductExternalTransformation(): ModelTransformation[] {
+	const transformations = [
+		new PropertyTypeTransformation(
+			{
+				buttonText: PropertyType.String,
+				externalUrl: PropertyType.String,
+			},
+		),
+		new KeyChangeTransformation< IProductExternal >(
+			{
+				buttonText: 'button_text',
+				externalUrl: 'external_url',
+			},
+		),
+	];
+
+	return transformations;
+}
+