+++ tests/e2e/specs/rest-api/external-product.js
@@ -1,0 +1,6 @@
+/*
+ * Internal dependencies
+ */
+const { runExternalProductAPITest } = require( '@woocommerce/e2e-core-tests' );
+
+runExternalProductAPITest();